+++ package.json
@@ -33,25 +33,25 @@
     "configuration": {
       "title": "Compare Method Version Settings",
       "properties": {
-        "compareMethodVersionSettings.methodSignature": {
-          "type": "string",
-          "default": "sub ([A-Za-z_]*)(\\d*)\\s",
-          "description": "Pattern used to find methods"
-        },
-        "compareMethodVersionSettings.methodVersion": {
-          "type": "string",
-          "default": "(.*sub [A-Za-z_]*)(\\d{4})(\\d{2})(\\d{2})(\\s.*)",
-          "description": "Pattern used to extract the version from a method's signature"
-        },
-        "compareMethodVersionSettings.methodVersionExtraction": {
-          "type": "string",
-          "default": "$1$5 # $3/$4/$2",
-          "description": "Template to refactor the method signature"
-        },
-        "compareMethodVersionSettings.methodDescription": {
-          "type": "string",
-          "default": "(^(?:#.*\\n)+)",
-          "description": "Pattern used to retrieve comments describing a method"
+        "compareMethodVersionSettings.methodPatterns": {
+          "type":"array",
+          "markdownDescription": "Patterns used to find and manipulate methods",
+          "default": [
+            {
+              "filetype": ".pm",
+              "signature": "sub ([A-Za-z_]*)(\\d*)\\s",
+              "version": "(.*sub [A-Za-z_]*)(\\d{4})(\\d{2})(\\d{2})(\\s.*)", 
+              "versionExtraction": "$1$5 # $3/$4/$2",
+              "description": "(^(?:#.*\\n)+)"
+            },
+            {
+              "filetype": ".t",
+              "signature": "subtest ([A-Za-z_]*)(\\d{4})(\\d{2})(\\d{2})\\s",
+              "version": "(.*subtest [A-Za-z_]*)(\\d{4})(\\d{2})(\\d{2})(\\s.*)",
+              "versionExtraction": "$1$5 # $3/$4/$2",
+              "description": "(^(?:#.*\\n)+)"
+            }
+          ]
         }
       }
     }
