+++ src/comparer/error-util.ts
@@ -1,0 +1,3 @@
+export const throwError = (errorMessage: string): never => {
+    throw new Error(errorMessage);
+};
