+++ src/comparer/regex-util.ts
